chores: presentation and documentation. (#400)
* chore: consistent import and export references.

* chore: comments about Flutter map properties and class.

* chore: comments about MapController and cleaner indentation.

* chore: remove redundant import to flutter/material.dart.

* chore: comments about MapOptions.

* chore: deprecated MapOptions.debug, dartfmt on tile_layer and
bonus comment for MapOptions and MapPosition.

* chore: move TileProvider in their own folder
