+++ lib/plugin_api.dart
@@ -1,6 +1,6 @@
 library flutter_map.plugin_api;
 
-export 'flutter_map.dart';
-export 'src/core/bounds.dart';
-export 'src/core/center_zoom.dart';
-export 'src/map/map.dart';
+export 'package:flutter_map/flutter_map.dart';
+export 'package:flutter_map/src/core/bounds.dart';
+export 'package:flutter_map/src/core/center_zoom.dart';
+export 'package:flutter_map/src/map/map.dart';
